requirements.txt
asgiref==3.7.1
distlib==0.3.6
Django==4.2.1
django-autoslug==1.9.9
django-coupon-management==1.2
django-jazzmin==2.6.0
django-jazzmin-admin-rangefilter==1.0.0
django-tinymce==3.6.1
#django-user-management==18.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
incuna-mail==4.1.1
incuna-pigeon==0.1.0
mysqlclient==2.1.1
Pillow==9.5.0
platformdirs==3.5.1
PyJWT==2.7.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
users==1.0.dev0
virtualenv==20.23.0
virtualenvwrapper-win==1.2.7
drf-yasg==1.21.6
django-clearcache==1.2.1